+++ tests/RequestTest.php
@@ -10,8 +10,8 @@ use seregazhuk\PinterestBot\Api\Request;
 use seregazhuk\PinterestBot\Api\CurlAdapter;
 use seregazhuk\tests\helpers\ResponseHelper;
 use seregazhuk\tests\helpers\ReflectionHelper;
+use seregazhuk\PinterestBot\Api\Contracts\Http;
 use seregazhuk\PinterestBot\Helpers\CsrfHelper;
-use seregazhuk\PinterestBot\Contracts\Http;
 
 /**
  * Class RequestTest.